+++ src/constants/routes.js
@@ -6,6 +6,7 @@ export const FAVOURITES_STACK = "FavouritesStack";
 export const ITCHIO_STACK = "ItchioStack";
 export const ITCHIO_AUTH_ROUTE = "ItchioAuthRoute";
 export const ITCHIO_STORE_ROUTE = "ItchioStoreRoute";
+export const ITCHIO_OWNED_ROUTE = "ItchioOwnedRoute";
 
 // Playdate
 export const PLAYDATE_STACK = "PlaydateStack";