+++ src/navigation/service.js
@@ -1,0 +1,6 @@
+import React from "react";
+export const navigationRef = React.createRef();
+
+export const useNavigationRef = () => {
+  return navigationRef;
+};