+++ src/navigation/stacks/mainStack/MainStack.jsx
@@ -1,10 +1,14 @@
 import {createStackNavigator} from "@react-navigation/stack";
 import React from "react";
+import useItchioStore from "../../../store/itchio";
+import itchioStackRoutes from "../itchioStack/ItchioStackRoutes";
 import mainStackRoutes from "./mainStackRoutes";
 
 const Stack = createStackNavigator();
 
 const MainStack = () => {
+  const {token} = useItchioStore();
+
   return (
     <>
       <Stack.Navigator
@@ -23,6 +27,15 @@ const MainStack = () => {
             options={item.options}
           />
         ))}
+        {token &&
+          itchioStackRoutes.map(item => (
+            <Stack.Screen
+              name={item.name}
+              component={item.component}
+              key={item.name}
+              options={item.options}
+            />
+          ))}
       </Stack.Navigator>
     </>
   );