+++ src/screens/Itchio/itchioOauth/ItchioOauth.jsx
@@ -1,9 +1,8 @@
-import React, {useCallback, useEffect, useState} from "react";
+import React, {useState} from "react";
 import {TextInput} from "react-native";
 import useItchioStore from "../../../store/itchio";
 import {Button, Text, View} from "react-native-windows";
 import {signInAsync} from "../../../helper/auth";
-import AsyncStorage from "@react-native-async-storage/async-storage";
 import BaseScreen from "../../../components/baseScreen/BaseScreen";
 
 const ItchioOAuth = () => {
@@ -15,15 +14,6 @@ const ItchioOAuth = () => {
     signInAsync();
     setAwait(true);
   };
-
-  const asyncLogin = useCallback(async () => {
-    const accessToken = await AsyncStorage.getItem("userToken");
-    validateToken(accessToken);
-  }, []);
-
-  useEffect(() => {
-    asyncLogin();
-  }, [asyncLogin]);
 
   const Success = () => (
     <View